+++ e2e_testing/torchscript/basic.py
@@ -586,144 +586,6 @@ def ExpandModule_basic(module, tu: TestUtils):
 
 # ==============================================================================
 
-
-class OnesModuleInt(torch.nn.Module):
-    def __init__(self):
-        super().__init__()
-
-    @export
-    @annotate_args([
-        None,
-    ])
-    def forward(self):
-        return torch.ones(3, 4, dtype=torch.int64)
-
-@register_test_case(module_factory=lambda: OnesModuleInt())
-def OnesModuleInt_basic(module, tu: TestUtils):
-    module.forward()
-
-# ==============================================================================
-
-class OnesModuleFloat(torch.nn.Module):
-    def __init__(self):
-        super().__init__()
-
-    @export
-    @annotate_args([
-        None,
-    ])
-    def forward(self):
-        return torch.ones(3, 4, dtype=torch.float32)
-
-@register_test_case(module_factory=lambda: OnesModuleFloat())
-def OnesModuleFloat_basic(module, tu: TestUtils):
-    module.forward()
-
-
-class OnesModuleFalsePinMemory(torch.nn.Module):
-    def __init__(self):
-        super().__init__()
-
-    @export
-    @annotate_args([
-        None,
-    ])
-    def forward(self):
-        return torch.ones(3, 4, dtype=torch.float32, pin_memory=False)
-
-@register_test_case(module_factory=lambda: OnesModuleFalsePinMemory())
-def OnesModuleFalsePinMemory_basic(module, tu: TestUtils):
-    module.forward()
-
-# ==============================================================================
-
-class EmptyIntModule(torch.nn.Module):
-    def __init__(self):
-        super().__init__()
-
-    @export
-    @annotate_args([
-        None,
-    ])
-    def forward(self):
-        return 0 * torch.empty((3, 4), dtype=torch.int64)
-
-@register_test_case(module_factory=lambda: EmptyIntModule())
-def EmptyModule_int(module, tu: TestUtils):
-    module.forward()
-
-# ==============================================================================
-
-class EmptyFloatModule(torch.nn.Module):
-    def __init__(self):
-        super().__init__()
-
-    @export
-    @annotate_args([
-        None,
-    ])
-    def forward(self):
-        return torch.pow(torch.empty((3, 4), dtype=torch.float32), 0)
-
-@register_test_case(module_factory=lambda: EmptyFloatModule())
-def EmptyModule_float(module, tu: TestUtils):
-    module.forward()
-
-
-class EmptyFalsePinMemoryModule(torch.nn.Module):
-    def __init__(self):
-        super().__init__()
-
-    @export
-    @annotate_args([
-        None,
-    ])
-    def forward(self):
-        return torch.pow(torch.empty((3, 4), dtype=torch.float32, 
-                                     pin_memory=False), 0)
-
-@register_test_case(module_factory=lambda: EmptyFalsePinMemoryModule())
-def EmptyModule_falsePinMemory(module, tu: TestUtils):
-    module.forward()
-
-# ==============================================================================
-
-class EmptyLikeIntModule(torch.nn.Module):
-    def __init__(self):
-        super().__init__()
-
-    @export
-    @annotate_args([
-        None,
-        ([-1, -1], torch.int64, True),
-    ])
-    def forward(self, a):
-        return 0 * torch.empty_like(a, dtype=torch.int64)
-
-@register_test_case(module_factory=lambda: EmptyLikeIntModule())
-def EmptyLikeModule_int(module, tu: TestUtils):
-    module.forward(torch.randint(10, (3, 5)))
-
-# ==============================================================================
-
-class EmptyLikeFloatModule(torch.nn.Module):
-    def __init__(self):
-        super().__init__()
-
-    @export
-    @annotate_args([
-        None,
-        ([-1, -1], torch.float32, True),
-    ])
-    def forward(self, a):
-        return torch.pow(torch.empty_like(a, dtype=torch.float32), 0)
-
-@register_test_case(module_factory=lambda: EmptyLikeFloatModule())
-def EmptyLikeModule_float(module, tu: TestUtils):
-    module.forward(tu.rand(4, 5))
-
-# ==============================================================================
-
 class ContiguousModule(torch.nn.Module):
     def __init__(self):
         super().__init__()
@@ -926,57 +788,6 @@ def DropoutModule_basic(module, tu: TestUtils):
     module.forward(tu.rand(3, 4))
 
 
-class Fill_TensorFloat64WithFloat32(torch.nn.Module):
-    def __init__(self):
-        super().__init__()
-
-    @export
-    @annotate_args([
-        None,
-        ([-1, -1, -1], torch.float32, True),
-    ])
-    def forward(self, tensor):
-        return torch.ops.aten.fill_(tensor, 3.0)
-
-@register_test_case(module_factory=lambda: Fill_TensorFloat64WithFloat32())
-def Fill_TensorFloat64WithFloat32_basic(module, tu: TestUtils):
-    module.forward(torch.randn(3, 2, 4))
-
-
-class Fill_TensorFloat64WithFloat64(torch.nn.Module):
-    def __init__(self):
-        super().__init__()
-
-    @export
-    @annotate_args([
-        None,
-        ([-1, -1, -1], torch.float64, True),
-    ])
-    def forward(self, tensor):
-        return torch.ops.aten.fill_(tensor, 3.0)
-
-@register_test_case(module_factory=lambda: Fill_TensorFloat64WithFloat64())
-def Fill_TensorFloat64WithFloat64_basic(module, tu: TestUtils):
-    module.forward(torch.randn(3, 2, 4).to(torch.float64))
-
-
-class Fill_TensorFloat64WithInt64(torch.nn.Module):
-    def __init__(self):
-        super().__init__()
-
-    @export
-    @annotate_args([
-        None,
-        ([-1, -1, -1], torch.float64, True),
-    ])
-    def forward(self, tensor):
-        return torch.ops.aten.fill_(tensor, 3)
-
-@register_test_case(module_factory=lambda: Fill_TensorFloat64WithInt64())
-def Fill_TensorFloat64WithInt64_basic(module, tu: TestUtils):
-    module.forward(torch.randn(3, 2, 4).to(torch.float64))
-
-
 class MeanModule(torch.nn.Module):
     def __init__(self):
         super().__init__()
@@ -1047,86 +858,6 @@ def NumelZeroRankModule_basic(module, tu: TestUtils):
     module.forward(torch.randint(10,[]))
 
 
-class ZerosModuleInt2D(torch.nn.Module):
-    def __init__(self):
-        super().__init__()
-
-    @export
-    @annotate_args([
-        None,
-    ])
-    def forward(self):
-        return torch.zeros(3, 4, dtype=torch.int64)
-
-@register_test_case(module_factory=lambda: ZerosModuleInt2D())
-def ZerosModuleInt2D_basic(module, tu: TestUtils):
-    module.forward()
-
-
-class ZerosModuleInt3D(torch.nn.Module):
-    def __init__(self):
-        super().__init__()
-
-    @export
-    @annotate_args([
-        None,
-    ])
-    def forward(self):
-        return torch.zeros(3, 4, 5, dtype=torch.int64)
-
-@register_test_case(module_factory=lambda: ZerosModuleInt3D())
-def ZerosModuleInt3D_basic(module, tu: TestUtils):
-    module.forward()
-
-
-class ZerosModuleFloat2D(torch.nn.Module):
-    def __init__(self):
-        super().__init__()
-
-    @export
-    @annotate_args([
-        None,
-    ])
-    def forward(self):
-        return torch.zeros(3, 4, dtype=torch.float32)
-
-@register_test_case(module_factory=lambda: ZerosModuleFloat2D())
-def ZerosModuleFloat2D_basic(module, tu: TestUtils):
-    module.forward()
-
-
-class ZerosModuleFloat3D(torch.nn.Module):
-    def __init__(self):
-        super().__init__()
-
-    @export
-    @annotate_args([
-        None,
-    ])
-    def forward(self):
-        return torch.zeros(3, 4, 5, dtype=torch.float32)
-
-@register_test_case(module_factory=lambda: ZerosModuleFloat3D())
-def ZerosModuleFloat3D_basic(module, tu: TestUtils):
-    module.forward()
-
-
-class ZerosModuleFalsePinMemory(torch.nn.Module):
-    def __init__(self):
-        super().__init__()
-
-    @export
-    @annotate_args([
-        None,
-    ])
-    def forward(self):
-        return torch.zeros(3, 4, dtype=torch.float32, pin_memory=False)
-
-@register_test_case(module_factory=lambda: ZerosModuleFalsePinMemory())
-def ZerosModuleFalsePinMemory_basic(module, tu: TestUtils):
-    module.forward()
-
-
 class BoolTensorReturnFalseModule(torch.nn.Module):
     def __init__(self):
         super().__init__()
@@ -1181,6 +912,7 @@ def BoolTensorReturnMixedModule_basic(module, tu: TestUtils):
     module.forward(torch.tensor([[1, 0], [0,1]], dtype=torch.bool))
 
 # ==============================================================================
+
 class TModuleRank2(torch.nn.Module):
     def __init__(self):
         super().__init__()
@@ -1193,10 +925,10 @@ class TModuleRank2(torch.nn.Module):
     def forward(self, lhs):
         return torch.t(lhs)
 
-
 @register_test_case(module_factory=lambda: TModuleRank2())
 def TModuleRank2_basic(module, tu: TestUtils):
     module.forward(tu.rand(3, 4))
+
 
 class TModuleRank1(torch.nn.Module):
     def __init__(self):
@@ -1210,10 +942,10 @@ class TModuleRank1(torch.nn.Module):
     def forward(self, lhs):
         return torch.t(lhs)
 
-
 @register_test_case(module_factory=lambda: TModuleRank1())
 def TModuleRank1_basic(module, tu: TestUtils):
     module.forward(tu.rand(3))
+
 
 class TModuleRank0(torch.nn.Module):
     def __init__(self):
@@ -1226,7 +958,6 @@ class TModuleRank0(torch.nn.Module):
     ])
     def forward(self, lhs):
         return torch.t(lhs)
-
 
 @register_test_case(module_factory=lambda: TModuleRank0())
 def TModuleRank0_basic(module, tu: TestUtils):
